+++ src/sjf.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
-#include <stdlib.h>
+#include <inttypes.h>
 #include "include/scheduling.h"
 #include "include/heap.h"
 
 void SJF(ProcessList *pl)
 {
-    size_t time = 0;
-    size_t total_wait = 0;
+    uint64_t time = 0;
+    uint64_t total_wait = 0;
     ProcessHeap *arrival_queue = createHeap(pl->size, ARRIVAL);
     ProcessHeap *ready_queue = createHeap(pl->size, BURST);
